Apply vanitygaps patch
This patch reimplements all the layouts (now saved on layouts.c) to add
gaps. This patch differentiates between inner and outer gaps as well as
horizontal and vertical gaps.

The logic of these layouts also aims to be pixel perfect by ensuring an
even split of the available space and re-distributing the remainder
among the available clients.